+++ elements/itemfilter/src/helpers/filter-external.js
@@ -1,3 +1,4 @@
+import { getValue } from "./";
 /**
  * Fetches and filters items based on the provided filters and configuration using an external API.
  *
@@ -7,12 +8,20 @@
  * @returns {Promise<Array<Object>>} The filtered items from the external API.
  */
 async function filterExternal(items, filters, config) {
+  const functionResult = config.externalFilter(items, filters);
+  // Check if the function returns a string or an object
+  // If string, assume this is the url to be fetched;
+  // If object, assume structure { url: "<API endpoint>", key?: "<nested.property.key for items>"}
+  const url =
+    typeof functionResult === "string" || functionResult instanceof String
+      ? functionResult
+      : functionResult.url;
   // Generate the URL for the external API based on the provided items and filters
-  const response = await fetch(`${config.externalFilter(items, filters)}`);
+  const response = await fetch(url);
   const jsonData = await response.json();
 
   // Return the array from the parsed JSON data
-  return jsonData;
+  return functionResult.key ? getValue(functionResult.key, jsonData) : jsonData;
 }
 
 export default filterExternal;