+++ elements/layercontrol/src/components/layer-list.js
@@ -24,6 +24,7 @@ export class EOxLayerControlLayerList extends LitElement {
     noShadow: { type: Boolean },
     toolsAsList: { type: Boolean },
     globallyExclusiveLayers: { type: Boolean },
+    customEditorInterfaces: { attribute: false, type: Array },
   };
 
   constructor() {
@@ -100,6 +101,15 @@ export class EOxLayerControlLayerList extends LitElement {
      * @type {Boolean}
      */
     this.globallyExclusiveLayers = false;
+
+    /**
+     * List of custom editor interfaces for layer config eox-jsonform
+     * Read more about the implementation of custom editor interfaces here:
+     * https://github.com/json-editor/json-editor/blob/master/docs/custom-editor.html
+     *
+     * @type {Array}
+     */
+    this.customEditorInterfaces = [];
   }
 
   /**
@@ -160,6 +170,8 @@ export class EOxLayerControlLayerList extends LitElement {
                             .toolsAsList=${this.toolsAsList}
                             .globallyExclusiveLayers=${this
                               .globallyExclusiveLayers}
+                            .customEditorInterfaces=${this
+                              .customEditorInterfaces}
                             @changed=${() => this.requestUpdate()}
                           >
                           </eox-layercontrol-layer-group>
@@ -177,6 +189,8 @@ export class EOxLayerControlLayerList extends LitElement {
                             .toolsAsList=${this.toolsAsList}
                             .globallyExclusiveLayers=${this
                               .globallyExclusiveLayers}
+                            .customEditorInterfaces=${this
+                              .customEditorInterfaces}
                             @changed=${() => this.requestUpdate()}
                           ></eox-layercontrol-layer>
                         `