+++ elements/layercontrol/src/components/layer-tools.js
@@ -34,6 +34,7 @@ export class EOxLayerControlLayerTools extends LitElement {
     noShadow: { type: Boolean },
     toolsAsList: { type: Boolean },
     embedded: { state: true },
+    customEditorInterfaces: { attribute: false, type: Array },
   };
 
   constructor() {
@@ -83,6 +84,15 @@ export class EOxLayerControlLayerTools extends LitElement {
     setTimeout(() => {
       this.embedded = this.parentElement?.tagName === "EOX-LAYERCONTROL-LAYER";
     });
+
+    /**
+     * List of custom editor interfaces for layer config eox-jsonform
+     * Read more about the implementation of custom editor interfaces here:
+     * https://github.com/json-editor/json-editor/blob/master/docs/custom-editor.html
+     *
+     * @type {Array}
+     */
+    this.customEditorInterfaces = [];
   }
 
   /**
@@ -90,6 +100,22 @@ export class EOxLayerControlLayerTools extends LitElement {
    */
   createRenderRoot() {
     return this.noShadow ? this : super.createRenderRoot();
+  }
+
+  /**
+   * Bubbles up the layerConfig:change event with the updated jsonform value and layer reference.
+   * @param {CustomEvent} evt
+   */
+  #handleLayerConfigChange(evt) {
+    this.dispatchEvent(
+      new CustomEvent("layerConfig:change", {
+        bubbles: true,
+        detail: {
+          jsonformValue: evt.detail.jsonformValue,
+          layer: evt.detail.layer,
+        },
+      }),
+    );
   }
 
   // Initializes '_removeButton' invoking 'removeButton' function with 'this' context.
@@ -141,7 +167,9 @@ export class EOxLayerControlLayerTools extends LitElement {
               .noShadow=${true}
               .layerConfig=${this.layer.get("layerConfig")}
               .unstyled=${this.unstyled}
+              .customEditorInterfaces=${this.customEditorInterfaces}
               @changed=${() => this.requestUpdate()}
+              @layerConfig:change=${this.#handleLayerConfigChange}
             ></eox-layercontrol-layerconfig>
           `,
         )}